(jblog3) 삭제 - 포스트 삭제시 userId, postNo 검사, 카테고리 삭제시 userId, categoryId 검사

+++ jblog3/src/main/java/com/poscodx/jblog/service/CategoryService.java
@@ -31,7 +31,9 @@ public class CategoryService {
 	
 	@Transactional
 	public void delete(Long no, String id) {
-		// post 삭제 전에 no, id 확인이 필요해보임..
+		if(categoryRepository.findByNoAndId(no, id) == null) {
+			return;
+		}
 		postRepository.deleteByCategoryNo(no);
 		categoryRepository.deleteByNoAndId(no, id);
 	}
