+++ jblog3/src/main/java/com/poscodx/jblog/service/PostService.java
@@ -2,13 +2,18 @@ package com.poscodx.jblog.service;
 
 import java.util.List;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.poscodx.jblog.repository.CategoryRepository;
 import com.poscodx.jblog.repository.PostRepository;
 import com.poscodx.jblog.vo.PostVo;
 
 @Service
 public class PostService {
+	@Autowired
+	private CategoryRepository categoryRepository;
+	
 	private PostRepository postRepository;
 
 	public PostService(PostRepository postRepository) {
@@ -27,7 +32,10 @@ public class PostService {
 		return postRepository.findPostByNo(no);
 	}
 	
-	public void delete(Long no) {
+	public void delete(Long no, String id) {
+		if(categoryRepository.findByPostNoAndId(no, id) == null) {
+			return;
+		}
 		postRepository.deleteByNo(no);
 	}
 }
